<refactor>: 중복 사용 이미지 분리
